# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Demo API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/ping": {
      "get": {
        "tags": [
          "ping"
        ],
        "summary": "Ping",
        "operationId": "ping_api_ping_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/todos": {
      "get": {
        "tags": [
          "todos"
        ],
        "summary": "Get Todos",
        "description": "Get all todo items, optionally filtered by assignment.",
        "operationId": "get_todos_api_todos_get",
        "parameters": [
          {
            "name": "assigned_to_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by assigned person ID",
              "title": "Assigned To Id"
            },
            "description": "Filter by assigned person ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoItemResponse"
                  },
                  "title": "Response Get Todos Api Todos Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "todos"
        ],
        "summary": "Create Todo",
        "description": "Create a new todo item.",
        "operationId": "create_todo_api_todos_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{todo_id}/done": {
      "patch": {
        "tags": [
          "todos"
        ],
        "summary": "Mark Todo Done",
        "description": "Mark a todo item as done.",
        "operationId": "mark_todo_done_api_todos__todo_id__done_patch",
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{todo_id}/assign": {
      "patch": {
        "tags": [
          "todos"
        ],
        "summary": "Assign Todo",
        "description": "Assign or unassign a todo item to/from a person.",
        "operationId": "assign_todo_api_todos__todo_id__assign_patch",
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todo Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/todos/{todo_id}": {
      "delete": {
        "tags": [
          "todos"
        ],
        "summary": "Delete Todo",
        "description": "Delete a todo item.",
        "operationId": "delete_todo_api_todos__todo_id__delete",
        "parameters": [
          {
            "name": "todo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Todo Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/people": {
      "get": {
        "tags": [
          "people"
        ],
        "summary": "Get People",
        "description": "Get all people.",
        "operationId": "get_people_api_people_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PersonResponse"
                  },
                  "type": "array",
                  "title": "Response Get People Api People Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "people"
        ],
        "summary": "Create Person",
        "description": "Create a new person.",
        "operationId": "create_person_api_people_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/people/{person_id}": {
      "patch": {
        "tags": [
          "people"
        ],
        "summary": "Update Person",
        "description": "Update a person's name.",
        "operationId": "update_person_api_people__person_id__patch",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Person Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "people"
        ],
        "summary": "Delete Person",
        "description": "Delete a person.",
        "operationId": "delete_person_api_people__person_id__delete",
        "parameters": [
          {
            "name": "person_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Person Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PersonCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PersonCreate",
        "description": "Request model for creating a person."
      },
      "PersonResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "PersonResponse",
        "description": "Response model for a person."
      },
      "PersonUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PersonUpdate",
        "description": "Request model for updating a person."
      },
      "PingResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "PingResponse",
        "description": "Response model for ping endpoint."
      },
      "TodoItemCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TodoItemCreate",
        "description": "Request model for creating a todo item."
      },
      "TodoItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "is_done": {
            "type": "boolean",
            "title": "Is Done"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "is_done",
          "created_at",
          "assigned_to_id"
        ],
        "title": "TodoItemResponse",
        "description": "Response model for a todo item."
      },
      "TodoItemUpdate": {
        "properties": {
          "assigned_to_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To Id"
          }
        },
        "type": "object",
        "title": "TodoItemUpdate",
        "description": "Request model for updating a todo item's assignment."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}